+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - HCI: ENABLE_HCI_SERIALIZED_CONTROLLER_OPERATIONS serializes Inquiry, Remote Name Request and Create Connection operations
 - GAP: support extended advertising with ENABLE_LE_EXTENDED_ADVERTISING
+- GAP: gap_enable_link_watchdog() uses automatic flush timeout to disconnect if data cannot be sent
 - ATT DB: provide gatt_server_get_handle_range_for_service_with_uuid16 to find included service within handle range
 - GATT Service: Audio Input Control Service Server (AICS 1.0)
 - GATT Service: Volume Control Service Server (VCS 1.0)